pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "DirCompare"
version = "1.3.0"
description = "Compare two directories and determine which is more up to date"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "DirCompare Contributors" },
]
keywords = ["directory", "compare", "diff", "file", "sync"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Environment :: Other Environment",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Filesystems",
    "Topic :: Utilities",
]

[project.scripts]
dircompare = "DirCompare.__main__:main"

[project.urls]
Repository = "https://forgejo.darkhorseinfosec.com/username/DirCompare"

[project.optional-dependencies]
build = ["pyinstaller>=6.0", "pillow>=10.0"]
docs = ["mkdocs-material"]
dev = ["pytest", "ruff"]

[tool.setuptools.packages.find]
include = ["DirCompare*"]

[tool.pytest.ini_options]
testpaths = ["."]
